Conflicts:
	ProjectDAE/web/event_details.xhtml
	ProjectDAE/web/responsible_details.xhtml

+++ ProjectDAE/src/java/ejbs/SubjectBean.java
@@ -10,7 +10,6 @@ import entity.Subject;
 import exceptions.EntityAlreadyExistsException;
 import exceptions.MyConstraintViolationException;
 import exceptions.Utils;
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 import javax.ejb.EJBException;
@@ -31,7 +30,7 @@ public class SubjectBean {
     @PersistenceContext(unitName = "ProjectDAEPU")
     private EntityManager em;
 
-    public void createSubject(Long id, String name, int courseYear, String scholarYear)
+    public void createSubject(int id, String name, int courseYear, String scholarYear)
         throws EntityAlreadyExistsException, MyConstraintViolationException {
         try {
             if (em.find(Subject.class, id) != null) {
